keras/requirements.txt
h5py==2.8.0
scipy==1.1.0
matplotlib==2.2.2
seaborn==0.8.1
numpy==1.14.2
tensorflow===2.0.0
Keras==2.2.0
click==6.7
pytest==3.6.3
pyyaml>=4.2b1
